+++ binary.go
@@ -21,7 +21,7 @@ const (
 	DCBA
 	// BADC 32-bit words may be represented in mid-big-endian format
 	BADC
-	// BADC 32-bit words may be represented in mid-little-endian format
+	// CDAB 32-bit words may be represented in mid-little-endian format
 	CDAB
 )
 const (
@@ -41,9 +41,45 @@ const (
 	MidLittleEndian
 )
 
+// RegDataType defines how to inteprete registers
+type RegDataType int
+
+const (
+	// RegisterArray register array, ex: [12345, 23456, 5678]
+	RegisterArray RegDataType = iota
+	// HexString hexadecimal string, ex: "112C004F12345678"
+	HexString
+	// Scale linearly scale
+	Scale
+	// UInt16 uint16 array
+	UInt16
+	// Int16 int16 array
+	Int16
+	// UInt32 uint32 array
+	UInt32
+	// Int32 int32 array
+	Int32
+	// Float32 float32 array
+	Float32
+)
+
+// bitStringToUInt8s converts bits string to uint8 array.
+// Source: function code 15
+func bitStringToUInt8s(bitString string) ([]uint8, error) {
+	var result = []uint8{}
+	for _, v := range strings.Split(bitString, ",") {
+		i, err := strconv.ParseUint(v, 10, 8)
+		if err != nil {
+			return nil, err
+		}
+		result = append(result, uint8(i))
+	}
+	return result, nil
+}
+
 // DecimalStringToRegisters converts decimal string to uint16/words array in big endian order.
 // Limitation: leading space before comma is not allowed.
-// Source: upstream
+// Source: upstream.
 func DecimalStringToRegisters(decString string) ([]uint16, error) {
 	var result = []uint16{}
 	for _, v := range strings.Split(decString, ",") {
@@ -53,12 +89,11 @@ func DecimalStringToRegisters(decString string) ([]uint16, error) {
 		}
 		result = append(result, uint16(i))
 	}
-
 	return result, nil
 }
 
 // HexStringToRegisters converts hexadecimal string to uint16/words array in big endian order.
-// Source: upstream
+// Source: upstream.
 func HexStringToRegisters(hexString string) ([]uint16, error) {
 	bytes, err := hex.DecodeString(hexString)
 	if err != nil {
@@ -68,13 +103,13 @@ func HexStringToRegisters(hexString string) ([]uint16, error) {
 }
 
 // BytesToHexString converts bytes array to hexadecimal string.
-// Ex. 112C004F12345678
+// Example: 112C004F12345678
 func BytesToHexString(bytes []byte) string {
 	return hex.EncodeToString(bytes)
 }
 
 // RegistersToBytes converts registers/uint16 array to byte array in big endian order.
-// Source: downstream
+// Source: downstream.
 func RegistersToBytes(data []uint16) ([]byte, error) {
 	buf := new(bytes.Buffer)
 	for _, v := range data {
@@ -87,24 +122,24 @@ func RegistersToBytes(data []uint16) ([]byte, error) {
 	return buf.Bytes(), nil
 }
 
-// LinearScalingRegisters scales the registers linearly
+// LinearScalingRegisters scales the registers linearly.
 // Equation:
-// 	Let originLow, originHigh, targetLow, targetHigh as a, b, c, d accordingly
-// 	output = c + (d - c) * (input - a) / (b - a)
-func LinearScalingRegisters(data []uint16, originLow, originHigh, targetLow, targetHigh float64) []float64 {
+// 	Let originLow, originHigh, targetLow, targetHigh as a, b, c, d accordingly.
+// 	Output = c + (d - c) * (input - a) / (b - a)
+func LinearScalingRegisters(data []uint16, originLow, originHigh, targetLow, targetHigh float64) []float32 {
 	l := len(data)
 	low := math.Min(originLow, originHigh)
 	high := math.Max(originLow, originHigh)
-	result := make([]float64, l)
+	result := make([]float32, l)
 
 	for idx := 0; idx < l; idx++ {
-		result[idx] = targetLow + (targetHigh-targetLow)*(math.Min(math.Max(low, float64(data[idx])), high)-low)/(high-low)
+		result[idx] = float32(targetLow + (targetHigh-targetLow)*(math.Min(math.Max(low, float64(data[idx])), high)-low)/(high-low))
 	}
 	return result
 }
 
-// BytesToFloat32s converts byte array to float32 array in four endian orders.
-// i.e., BigEndian (0)
+// BytesToFloat32s converts byte array to float32 array in four endian orders. i.e.,
+//       BigEndian (0),
 //       LittleEndian (1)
 //       MidBigEndian (2)
 //       MidLittleEndian (3)
@@ -130,8 +165,8 @@ func BytesToFloat32s(buf []byte, endian Endian) ([]float32, error) {
 	return result, nil
 }
 
-// BytesToInt32s converts byte array to Int32 array in four endian orders.
-// i.e., BigEndian (0)
+// BytesToInt32s converts byte array to Int32 array in four endian orders. i.e.,
+//       BigEndian (0),
 //       LittleEndian (1)
 //       MidBigEndian (2)
 //       MidLittleEndian (3)
@@ -156,8 +191,8 @@ func BytesToInt32s(buf []byte, endian Endian) ([]int32, error) {
 	return result, nil
 }
 
-// BytesToUInt32s converts byte array to UInt32 array in four endian orders.
-// i.e., BigEndian (0)
+// BytesToUInt32s converts byte array to UInt32 array in four endian orders. i.e.,
+//       BigEndian (0),
 //       LittleEndian (1)
 //       MidBigEndian (2)
 //       MidLittleEndian (3)